+++ src/internal/errors.ts
@@ -1,3 +1,4 @@
+import { format } from 'util';
 import { allVirtualoptions } from './utils';
 
 export class InvalidPropError extends Error {
@@ -13,14 +14,32 @@ export class InvalidTypeError extends Error {
 }
 
 export class NotNumberTypeError extends Error {
-  constructor(key: string) {
-    super(`Type of ${key} property is not a number.`);
+  constructor(targetName: string, key: string, enumKey?: string, enumValue?: string) {
+    if (enumKey || enumValue) {
+      super(format(
+        'Typeof "%s.%s" is "Number", value is undefined/null or does not have a reverse mapping!\n'
+        + 'Encountered with property: %s.%s',
+        targetName, key,
+        enumKey, typeof enumValue
+      ));
+    } else {
+      super(`Type of "${targetName}.${key}" property is not a number.`);
+    }
   }
 }
 
 export class NotStringTypeError extends Error {
-  constructor(key: string) {
-    super(`Type of ${key} property is not a string.`);
+  constructor(targetName: string, key: string, enumKey?: string, enumValue?: string) {
+    if (enumKey || enumValue) {
+      super(format(
+        'Typeof "%s.%s" is "String", enum is not only Strings!\n'
+        + 'Encountered with property: %s.%s',
+        targetName, key,
+        enumKey, typeof enumValue
+      ));
+    } else {
+      super(`Type of "${targetName}.${key}" property is not a string.`);
+    }
   }
 }